Add files via upload
Abstract
As primary sources of energy are depleting day by day, so power generated by Non-Conventional sources such as photovoltaic (PV), wind turbines, fuel cell, small hydro etc. are very effective in overcoming the increasing demand. The main advantages of using renewable sources are the elimination of harmful emissions and the inexhaustible resources of the primary energy. However, the main disadvantage, apart from the higher costs, e.g., photovoltaic, is the uncontrollability. The availability of renewable sources has strong daily and seasonal patterns. Each renewable source has its own different characteristics. Therefore, it would be difficult to meet demands with a single renewable source. The way of fully exploiting the renewable energy is by operating two or more renewable sources in parallel. The objective of this project is to design and develop a suitable control strategy for non-conventional energy source fed dc dc converters operating in parallel. The non-conventional energy sources to be used are photovoltaic cell and fuel cell or battery. The control scheme should be such as to extract maximum power from the photovoltaic cell, and the rest of the demand should be met by the fuel cell or battery. In case of unavailability of power from the photovoltaic cell, whole of the power should be supplied by the fuel cell or battery.